bot/requirements.txt
marshmallow==3.15.0
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
